Cache station metadata for a day
Station metadata changes extremely rarely, if ever, so it seems awfully
silly to keep downloading the exact same metadata every time we extract.
Wastes requests, wastes time (300ms round trip for me)

This commit makes it so station metadata is cached for a day, and cached
data is used in place of downloaded metadata when available.

Closes #16