add support for https://radiko.jp/persons/ URLs
those urls give a bio and pic of the person, and recently they've
started giving a list of all the programmes that person's appeared in as
well.

the API is a bit annoying, you have to specify a start and end time to
search between, so i had to work out the logic for "everything currently
on timefree"

also, the site doesn't use the "end_at_lt" parameter, only start_at_lt
i tried it in the url on the offchance it would work and save a bit of
logic
it did so i will keep it, but it probably stands out a quite a bit, idk
if the mobile apps use this, or with the end_at_lt parameter

the only playlist details I can provide are the ID. There is an API for
the actual details, but that's an extra request, and it's protobuf.